added chrono capabilities based on distance
* 00..10 => beep immediately.
* 10..... => set chrono at (cm-10) minutes.

The problem is to store the flags so that a chrono is not set while
running a countdown, or set again when it finishes at wtv distance the
marker is.

A lot of flags that show I don't know what I'm doing. This looks like
it should be an FSM. Or use Ferret.